Refined several scripts, reworked leveling and damage, increased overall weapon speed and range, decreased base damage. Added victory screen. Changed timer to enemies remaining. New mechanics: Destroying enemies without taking damage adds to bonus damage multiplier. Enemy Managers spawn enemies faster if there are not many left.

+++ World of Edge/Assets/Scripts/Bullet.cs
@@ -21,7 +21,7 @@ public class Bullet : BulletParent
             {
                 enemy.Flash();
             }
-            health.takeDamage((int)damage);
+            health.takeDamage(damage);
             Destroy(gameObject);
         }
         else
